+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [5.3.0] - 2024-05-13
+
 ### Added
 
 - upcloud_network_peering: support for network peerings.
@@ -593,7 +595,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.2.3...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.3.0...HEAD
+[5.3.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.2.3...v5.3.0
 [5.2.3]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.2.2...v5.2.3
 [5.2.2]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.2.1...v5.2.2
 [5.2.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.2.0...v5.2.1